+++ appseed/ca/sockets/sockets_link_out_socket.cpp
@@ -46,7 +46,7 @@ namespace sockets
 
    }
 
-   void link_out_socket::link_write(void * p, int size)
+   void link_out_socket::link_write(void * p, size_t size)
    {
       m_out->SendBuf((const char *) p, size);
    }
